Fixed moving the first platform to the end
The problem was that the rigidBodyControl needs to be disabled, using
setEnabled(false), before it can be moved to prevent mysterious bugs.
Included get- and setPlatformPosition in the Platform class, and the
setter disables, moves then reenables the rbControl. The game now runs
forever, though you run out of light after a while.