Background walls
Added stone walls behind the platforms. This affects the generateChunk
method in LevelControl and the new class Wall (modelled on Platform but
without any physics). I also changed the numbers given to the Platform
so that the real measurements are in P (the Box constructor doubles the
input for some reason, so that a length of 2 means it measures 2 from
the middle to each edge => 4 in total).