Changed the current wall texture to the one in the image used in the presentation. Fixed so that the texture repeats over each wall chunk without being disorted for some screen resolutions (whenever the aspect ratio is not the same as with 640x480, note that the phone uses 800x480).
In order to do this, I introduced screenWidth and screenHeight variables in P.java, which are given values in Main.java. Wall.java then uses these variables to calculate how many times the texture should be repeated over the wall chunk. Note that this will not work if the size of the chunks are changed later on, but the expression can be changed to accommodate to that as well.