Added LevelChunk.activate()
Since the LevelChunk knows about the lights connected to it but doesn't
have them attached to itself, it has to know about its parent node
ahead of time. Thus it can also have a method for attaching itself and
all lights in one go, just like the pre-existing remove() which did the
opposite.