Built a Hazard framework
Made class Hazard abstract, and created specific subclasses for the
different types of fireballs. A specific fireball, like
HoveringFireballHazard, is now completely self-contained and the
ChunkFactory only needs to set any constructor-taken parameters and add
it to the chunk, just like platforms. (Previously it needed to also
create a FireballControl and connect them to each other.)