Game objects generating new game objects
Added the possibility for game objects (spatials with controls) to
generate another spatial, in the current only example a Wizard
generating Fireballs. Each window in the level now has a WizardControl
that knows about the LevelControl (LevelControl checks all level
content for anything that implements LevelContentGenerator to ensure
this). The Wizards then spawn Fireballs and give them to the
LevelControl by the new levelControl.addToLevel():