Removed the LevelChunk class
LevelControl now treats all spatials equally, and checks each one for
removal. No objects of importance need to be grouped together, so the
LevelChunk class is scrapped. Lights belonging to a single chunk is
currently not supported. The Node containing the background acts as a
checkpoint node – its destruction due to out-of-bounds triggers the
creation of a new chunk of the level.