Testing the mid-air dash
Testing the mid-air dash as a solution to unfair impossible deaths.
Player can now do a very small second jump in mid-air, and can do it
again if it gets hit by a fireball. This means it can start moving
forward in the air after a knockback, which gives the player control
over where he lands. My thoughts: solves the problem well but might not
be thematic. Not sure if dash should be allowed before being hit at all.