Removed dependency on jme3testdata
To fix the bloated apk size: stops importing entire jme3testdata library
(project.properties).
Separate import only of the particle effects that we want (all the png
files in this commit).
Also updated all the .java files that use a particle effect to look in
the new path for the image file.