Let player fall out of view after game over
After the game over menu is brought up, the player keeps falling and
stops when outside the camera view.
Also: following a tip that I'm not sure if it applies to us, but if it
does it might solve the strange nifty nullpointerexceptions. Please
report if you keep getting them.