Fixed bat rotation and testing random colors
Set the rotation for the bat so it faces the player
Testing random material colors to the bat. The colors are not intended
to be permanent, but it's good to see how this works, and also they need
better visibility with the current lighting settings.
Randomizing between different brown / grey / black / other colors can be
nice for the visual variety.