Bigger blender platforms.
added size 24 and 36 platforms. 40 would be a lot more work and there's
an advantage of 12n sizes, so I'm suggesting 36 instead of 40. Is that
possible?

Graphics can easily be changed in terms of color, blur, filter, etc, as
long as the brick pattern remains. The file loaded in the model is
untitled8.png, the other png's are variants